+++ src/ea/edu/Dreieck.java
@@ -7,7 +7,7 @@ import ea.internal.annotations.API;
 @API
 public class Dreieck extends EduGeometrie<Polygon> {
     @API
-    public Dreieck(float ax, float ay, float bx, float by, float cx, float cy) {
+    public Dreieck(double ax, double ay, double bx, double by, double cx, double cy) {
         super(new Polygon(new Vector(ax, ay), new Vector(bx, by), new Vector(cx, cy)));
     }
 }
